+++ src/lib/events/eventIfActorAtPosition.js
@@ -27,6 +27,7 @@ const fields = [
   },
   {
     type: "group",
+    wrapItems: true,
     fields: [
       {
         key: "x",